src/training/requirements.txt
webdatasets
torch
torchvision
ffcv
Pillow